To master (#23)
* Task 1: create add product controller (#1)

* Initial commit

* refactoring

* add package-info

* add package-info

* refactoring

* add swagger

* refactoring

* Task 2 : add h2 database and field article (#2)

* add h2 database and field article

* add validation on filed 'article'

* Task 3 : add GlobalHandlerExecption (#3)

* add GlobalExceptionHandler

* refactoring

* Task 4: add ProductPriceSheduler and pagination on findAll controller (#4)

* test commit

* add ProductPriceSheduler

* refactoring article migration

* add article on updateProductRequest

* refactoring

* refactoring

* refactoring

* add ExecutionTimeAspect (#5)

* Task 6: add ProductServiceImplTest (#6)

* add ProductServiceImplTest

* refactoring

* refactoring

* refactoring

* refactoring article migration

* Task 7: add ProductControllerTest (#8)

* add ProductControllerTest

* refactoring

* Task 8: add new endpoint searchProducts (#9)

* add new endpoint searchProducts

* refactoring

* refactoring

* add description on endpoint

* refactoring

* Task 9: add optimistic_read (#10)

* add optimistic locking

* refactoring

* refactoring

* Task 10: add saveProductsToXlsx and endpoint getReportFileNames (#11)

* Task 11: add endpoints upload/download file (#12)

* add endpoints upload/download file

* change media-type

* Task 12: add ExchangeRateProvider on getProductById (#13)

* add ExchangeRateProvider on getProductById

* refactoring

* refactoring

* refactoring

* refactoring

* add slf4j

* refactoring

* reafctoring

* refactoring

* fix article exists (#14)

* Task 13: add cache and 2 services (#15)

* add cache and 2 services

* refactoring

* Task 14: add ExchangeRateClientImplStub (#16)

* add cache and 2 services

* refactoring

* test

* test

* Task 15: add 3 currency for headers (#17)

* add 3 currency for headers

* refactoring

* refactoring

* Task 16 and 16 +1: add INN service and refactoring (#19)

* add services order and customer

* add getOrdersByProductId

* add INN service and refactoring

* Task 17: add event controller (#20)

* add event controller

* refactoring

* refactoring

* Task 17+1: add kafka (#21)

* add event controller

* refactoring

* refactoring

* add kafka

* Task 18: add redis (#22)

* add redis

* refactoring

* refactoring

+++ src/test/java/com/crudshop/demo/service/ProductServiceImplTest.java
@@ -8,6 +8,7 @@ import com.crudshop.demo.exception.ProductNotFoundException;
 import com.crudshop.demo.motherobject.ProductDtoBuilder;
 import com.crudshop.demo.motherobject.ProductEntityBuilder;
 import com.crudshop.demo.repository.ProductRepository;
+import com.crudshop.demo.service.product.ProductServiceImpl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -20,6 +21,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
@@ -131,7 +133,7 @@ public class ProductServiceImplTest {
                 .withName("Слива")
                 .withDescription("Зеленый")
                 .withCategories(Categories.VEGETABLES)
-                .withPrice(25.00)
+                .withPrice(BigDecimal.valueOf(25.00))
                 .withQuantity(25)
                 .build();
